Make things prettier

+++ contracts/Fund.sol
@@ -1,29 +1,42 @@
-import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
-import '../interfaces/IWETH.sol';
-import './RoleAware.sol';
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import "../interfaces/IWETH.sol";
+import "./RoleAware.sol";
 
 contract Fund is RoleAware {
     address public WETH;
     address[] public approvedTokens;
     mapping(address => bool) public activeTokens;
-    
+
     constructor(address _WETH, address _roles) RoleAware(_roles) {
         WETH = _WETH;
     }
 
-    function deposit(address depositToken, uint depositAmount) external returns (bool) {
+    function deposit(address depositToken, uint256 depositAmount)
+        external
+        returns (bool)
+    {
         require(activeTokens[depositToken], "Deposit token is not active");
-        return IERC20(depositToken).transferFrom(msg.sender,
-                                                 address(this),
-                                                 depositAmount);
+        return
+            IERC20(depositToken).transferFrom(
+                msg.sender,
+                address(this),
+                depositAmount
+            );
     }
 
-    function depositFor(address sender, address depositToken, uint depositAmount) external returns (bool) {
+    function depositFor(
+        address sender,
+        address depositToken,
+        uint256 depositAmount
+    ) external returns (bool) {
         require(activeTokens[depositToken], "Deposit token is not active");
         require(isWithdrawer(msg.sender), "Contract not authorized to deposit");
-        return IERC20(depositToken).transferFrom(sender,
-                                                 address(this),
-                                                 depositAmount);
+        return
+            IERC20(depositToken).transferFrom(
+                sender,
+                address(this),
+                depositAmount
+            );
     }
 
     function depositToWETH() external payable {
@@ -31,22 +44,31 @@ contract Fund is RoleAware {
     }
 
     // withdrawers role
-    function withdraw(address withdrawalToken, address recipient, uint withdrawalAmount)
-        external returns (bool) {
-        require(isWithdrawer(msg.sender), "Contract not authorized to withdraw");
+    function withdraw(
+        address withdrawalToken,
+        address recipient,
+        uint256 withdrawalAmount
+    ) external returns (bool) {
+        require(
+            isWithdrawer(msg.sender),
+            "Contract not authorized to withdraw"
+        );
         return IERC20(withdrawalToken).transfer(recipient, withdrawalAmount);
     }
 
     // withdrawers role
-    function withdrawETH(address recipient, uint withdrawalAmount)
-        external {
+    function withdrawETH(address recipient, uint256 withdrawalAmount) external {
         require(isWithdrawer(msg.sender), "Not authorized to withdraw");
         IWETH(WETH).withdraw(withdrawalAmount);
         payable(recipient).transfer(withdrawalAmount);
     }
 
     // withdrawers role
-    function sendTokenTo(address token, address recipient, uint amount) external returns (bool) {
+    function sendTokenTo(
+        address token,
+        address recipient,
+        uint256 amount
+    ) external returns (bool) {
         require(isWithdrawer(msg.sender), "Not authorized to withdraw");
         return IERC20(token).transfer(recipient, amount);
     }
